straver/requirements.txt
fastapi==0.110.0
uvicorn[standard]==0.27.1
psycopg2-binary==2.9.9
requests==2.31.0
pydantic==1.10.14
python-multipart==0.0.9
